+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LASRV011 - 04/04/2023 - Add the AuthUser class.
 - LASRV011 - 05/04/2023 - Add the RSAGenerator class.
 - LASRV011 - 05/04/2023 - Add the RSACommunicator class.
+- LASRV011 - 05/04/2023 - Add the PlainCommunicator class.
 
 ### Changed
 - LASRV010 - 03/04/2023 - Changes in Request class.